requirements.txt
Flask==0.12.2
line-bot-sdk==1.8.0
gunicorn==19.7.1
Keras==2.1.5
tensorflow==1.6.0
numpy==1.16.1
opencv-python==4.0.0.21
 pillow==5.4.1
